+++ server/routes/user.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import User from '../controllers/user';
+
+const router = express.Router();
+
+router.post('/api/v1/auth/signup', User.create);
+
+export default router;
